+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[5.0.4](https://github.com/multiversx/mx-sdk-dapp/pull/1516)] - 2025-07-16
+
+- [Fixed cancelling sign transactions/message for WalletConnect in UI](https://github.com/multiversx/mx-sdk-dapp/pull/1515)
+- [Added support for custom theme](https://github.com/multiversx/mx-sdk-dapp/pull/1514)
+- [Fixed init app double init](https://github.com/multiversx/mx-sdk-dapp/pull/1513)
+
 ## [[5.0.3](https://github.com/multiversx/mx-sdk-dapp/pull/1509)] - 2025-07-15
 
 - [Updated bundler to use esm](https://github.com/multiversx/mx-sdk-dapp/pull/1508)